+++ HISTORY.rst
@@ -8,6 +8,8 @@ Pending Release
 
 .. Insert new release notes below this line
 
+* Make ``public_*`` properties return ``None` for instances that aren't public.
+
 1.2.0 (2017-08-26)
 ------------------